+++ genesis-creator/src/genesis.rs
@@ -888,6 +888,10 @@ pub enum GenesisData {
         core: CoreGenesisParametersV1,
         initial_state: GenesisStateCPV3,
     },
+    P11 {
+        core: CoreGenesisParametersV1,
+        initial_state: GenesisStateCPV3,
+    },
 }
 
 impl GenesisData {
@@ -995,6 +999,15 @@ impl GenesisData {
                 core.serial(&mut hasher);
                 initial_state.serial(&mut hasher);
             }
+            GenesisData::P11 {
+                core,
+                initial_state,
+            } => {
+                // tag of initial genesis
+                0u8.serial(&mut hasher);
+                core.serial(&mut hasher);
+                initial_state.serial(&mut hasher);
+            }
         }
         let bytes: [u8; 32] = hasher.finalize().into();
         bytes.into()
@@ -1026,6 +1039,7 @@ pub fn make_genesis_data_cpv0(
         ProtocolVersion::P8 => None,
         ProtocolVersion::P9 => None,
         ProtocolVersion::P10 => None,
+        ProtocolVersion::P11 => None,
     }
 }
 
@@ -1128,6 +1142,16 @@ impl Serial for GenesisData {
                 initial_state,
             } => {
                 12u8.serial(out);
+                // tag of initial genesis
+                0u8.serial(out);
+                core.serial(out);
+                initial_state.serial(out)
+            }
+            GenesisData::P11 {
+                core,
+                initial_state,
+            } => {
+                13u8.serial(out);
                 // tag of initial genesis
                 0u8.serial(out);
                 core.serial(out);