+++ genesis-creator/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased changes
 
+- Support genesis data format of protocol version 11.
+
 ## 0.7.0
 
 - Support genesis data format of protocol version 10.